requirements.txt
opencv-python-headless==4.8.1.78
face-recognition==1.3.0
boto3==1.34.34
google-cloud-vision==3.4.5
plotly==5.17.0
seaborn==0.12.2
scipy==1.11.4
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
requests==2.31.0
Pillow==10.1.0
matplotlib==3.8.2
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.8.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
